+++ org.toxsoft.uskat.s5.histdata10/src/org/toxsoft/uskat/s5/server/backend/supports/histdata/impl/sequences/sync/S5HistDataSyncFloatingBlobEntity0.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
  */
 @Entity
 public class S5HistDataSyncFloatingBlobEntity0
-    extends S5SequenceSyncBlob<S5HistDataSyncFloatingEntity0, double[], double[]> {
+    extends S5SequenceSyncBlob<S5HistDataSyncFloatingEntity0, double[], byte[]> {
 
   private static final long serialVersionUID = 157157L;
