+++ org.toxsoft.uskat.s5.histdata10/src/org/toxsoft/uskat/s5/server/backend/supports/histdata/impl/sequences/sync/S5HistDataSyncIntegerBlobEntity9.java
@@ -15,7 +15,7 @@ import jakarta.persistence.*;
  */
 @Entity
 public class S5HistDataSyncIntegerBlobEntity9
-    extends S5SequenceSyncBlob<S5HistDataSyncIntegerEntity9, long[], long[]> {
+    extends S5SequenceSyncBlob<S5HistDataSyncIntegerEntity9, long[], byte[]> {
 
   private static final long serialVersionUID = 157157L;
